1）更新hibernate orm框架测试项目 2） 添加通用DaoTester 3）更新ctrip dao项目

+++ common-orm/src/main/java/com/pphh/demo/dao/EmployeeDao.java
@@ -40,7 +40,7 @@ public interface EmployeeDao {
      * @param employee a employee entity to be inserted
      * @return the id of new employee entity
      */
-    public int insert(EmployeeEntity employee);
+    public long insert(EmployeeEntity employee);
 
     /**
      * update an employee entity
